+++ mystash-service/src/main/java/org/ranjangeorge/mystash/service/impl/ledger/FetchBalance.java
@@ -7,19 +7,17 @@ import org.ranjangeorge.mystash.service.api.support.UsecaseNames;
 import org.ranjangeorge.mystash.service.impl.Stash;
 import org.ranjangeorge.mystash.service.impl.support.db.SessionFactoryHolder;
 
-import java.math.BigDecimal;
-
 @UsecaseNames(Usecase.FETCH_BALANCE)
 public class FetchBalance {
 
-    public BigDecimal fetchBalance(String stashId) {
+    public long fetchBalance(String stashId) {
 
         Session session = SessionFactoryHolder.getCurrentSession();
         Transaction transaction = session.beginTransaction();
 
         try {
 
-            BigDecimal balance = session.load(Stash.class, stashId).getBalance();
+            long balance = session.load(Stash.class, stashId).getBalance();
 
             // Commit
             transaction.commit();
